fixes order men to have name
