+++ server/controllers/Orders.js
@@ -65,7 +65,7 @@ class Orders {
   }
 
   /**
-   * This function get all orders
+   * This function get a specific order
    * @param {object} req - the request file.
    * @param {object} res - The response file.
    * @returns {object} Returns an order information.
@@ -95,40 +95,9 @@ class Orders {
         }));
   }
 
-  /**
-   * This function get the order history of a user
-   * @param {object} req - the request file.
-   * @param {object} res - The response file.
-   * @returns {object} Returns the order information.
-   */
-  userGetAOrderHistory(req, res) {
-    ordersModel.userGetAOrderHistory(req.params.id)
-      .then((result) => {
-        if (result.rowCount === 0) {
-          return res.status(200)
-            .json({
-              status: 'success',
-              message: 'No Order History Available for this User',
-              order: []
-            });
-        }
-        return res.status(200)
-          .json({
-            status: 'success',
-            message: 'Order History Successful',
-            order: result.rows
-          });
-      })
-      .catch(() => res.status(500)
-        .json({
-          status: 'Failed',
-          error: 'Failed to load Order'
-        }));
-  }
-
 
   /**
-   * This function get the order history of a user
+   * This function edit the order status of an order
    * @param {object} req - the request file.
    * @param {object} res - The response file.
    * @returns {object} Returns the order information.
