+++ server/models/OrdersModel.js
@@ -2,7 +2,7 @@
 import DbConnect from './DbConnect';
 
 /**
- * Represents the meal/menu model.
+ * Represents the orders model.
  */
 class OrdersModel extends DbConnect {
   /**
@@ -54,19 +54,6 @@ class OrdersModel extends DbConnect {
        */
   getASpecificOrders(data) {
     const sql = 'SELECT * FROM orders WHERE order_id = $1';
-    const param = [data];
-    return this.pool.query(sql, param);
-  }
-
-  /**
-       * This function get all order by a user
-       * @param {object} data - the req.params object .
-       * @returns {Promise} Returns the queried data .
-       */
-  userGetAOrderHistory(data) {
-    const sql = `SELECT * FROM orders 
-      WHERE user_id = $1 
-      ORDER BY order_id DESC`;
     const param = [data];
     return this.pool.query(sql, param);
   }
